refactor: Move BarTheme to ducks/bar
