+++ src/ducks/bar/BarTheme.jsx
@@ -1,7 +1,21 @@
+/* global cozy, __TARGET__ */
+
 import React from 'react'
 import PropTypes from 'prop-types'
 import withSideEffect from 'react-side-effect'
-import { setBarTheme } from 'ducks/mobile/utils'
+import { setTheme as setStatusBarTheme } from './statusBar'
+import barOverrides from 'ducks/bar/overrides'
+
+export const setBarTheme = theme => {
+  if (__TARGET__ === 'mobile') {
+    setStatusBarTheme(theme)
+  }
+
+  if (cozy.bar && cozy.bar.setTheme) {
+    const overrides = barOverrides[theme]
+    cozy.bar.setTheme(theme, overrides)
+  }
+}
 
 class BarTheme extends React.Component {
   render() {
