+++ src/ducks/bar/statusBar.js
