+++ src/ducks/pin/PinWrapper.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import LockedBody from 'ducks/pin/LockedBody'
 import styles from 'ducks/pin/styles.styl'
-import BarTheme from 'ducks/mobile/BarTheme'
+import BarTheme from 'ducks/bar/BarTheme'
 import Portal from 'cozy-ui/react/Portal'
 import cx from 'classnames'
